get.primes {randtoolbox}	R Documentation

Get primes for quasi random number generation

Description

Provides a vector of a specified number of smallest primes from the internal table of the package.

Usage

get.primes(n)

Arguments

n
The required number of primes. Should be at most 100 000.

Details

The package contains an internal table of the smallest 100 000 primes, which may be used in torus algorithm.

Value

Vector of min(n, 100000) smallest primes.

See Also

torus

Examples

p <- get.primes(20)
torus(5,dim=10,prime=p[11:20])

[Package randtoolbox version 2.0.5 Index]